refactor: renamed filename `similar_to.rs`
...to `similar_to_escape.rs` to avoid ambiguity.
