+++ src/expr/impl/src/scalar/mod.rs
@@ -68,7 +68,7 @@ mod repeat;
 mod replace;
 mod round;
 mod sha;
-mod similar_to;
+mod similar_to_escape;
 mod split_part;
 mod string;
 mod string_to_array;
