+++ src/expr/impl/src/scalar/similar_to_escape.rs
